Separate camelize and classify lookup
Previously, if you were using ActiveSupport, the lookup would use Rails'
`String#classify` method. Without ActiveSupport it'd use a patched
version which implemented classifying as camelizing a String.

As an example, if the lookup was given a symbol `:users`, it'd behave as
follows:
a) Without ActiveSupport, look for `UsersPolicy`;
b) With ActiveSupport, look for `UserPolicy`;

This change does three things. First, we add a new lookup that uses
`camelize`. This enables the lookup to find `UsersPolicy` when given
`:users` in Rails. This is the new behaviour.

Second, by changing the Symbol extension, we keep the old behaviour in
the non-Rails working.

Finally, we add a new lookup that is only used when `classify` is
available on String and uses that to infer the polict. This ensures that
finding `UserPolicy` still works in the case where we're in a Rails
application.