Fix global policies not being found with `strict_namespace: true`
Previously, we'd skip some lookups when `namespace` was nil and others
when `strict` was true. This resulted in a situation where you'd not
find a non-namespaced `ThingPolicy` when using the lookup with
`strict_namespace: true`.

With the changes of this commit, the lookup will start finding
`ThingPolicy` in strict mode if and only if the authorisation namespace
is nil.