Support to assert contexts on test matchers
On a project that I work on, we are using a lot the feature of passing
more context to the policies, while searching the documentation and the
source code I found that currently there's no way to assert the context.

This commit implements a way to assert what context is being passed to
the policies, on both `be_authorized` matcher for RSpec and
`assert_authorized_to` on Minitest.